Add async authentication check to component will mount in portfolio component
Need to update check to a function that can be accessible from all components. Need to decide how to execute this. For now it can go in every component that needs authentication. Later, it should go in a wrapper where it only happens once and passes to the appropriate route/component.